feat: separated API for persistence
