+++ backend/businesslogic/pom.xml
@@ -10,13 +10,13 @@
     <dependencies>
         <dependency>
             <groupId>io.github.fontysvenlo.ais</groupId>
-            <artifactId>persistence</artifactId>
+            <artifactId>datarecords</artifactId>
             <version>1.0-SNAPSHOT</version>
         </dependency>
         <dependency>
-            <groupId>io.github.fontysvenlo.ais</groupId>
-            <artifactId>datarecords</artifactId>
-            <version>1.0-SNAPSHOT</version>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>persistence-api</artifactId>
+            <version>${project.version}</version>
         </dependency>
     </dependencies>
 </project>
