+++ backend/businesslogic/src/main/java/module-info.java
@@ -3,7 +3,7 @@
  */
 module businesslogic_module {   
     requires datarecords_module;
-    requires persistence_module;
+    requires persistence_api_module;
     
     exports io.github.fontysvenlo.ais.businesslogic;
 }
