+++ backend/persistence-api/pom.xml
@@ -6,30 +6,19 @@
         <artifactId>airlineinformationsystem</artifactId>
         <version>1.0-SNAPSHOT</version>
     </parent>
-    <artifactId>persistence</artifactId>
+    <artifactId>persistence-api</artifactId>
+    <packaging>jar</packaging>
     <dependencies>
         <dependency>
-            <groupId>io.github.fontysvenlo.ais</groupId>
-            <artifactId>datarecords</artifactId>
-            <version>1.0-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>org.postgresql</groupId>
-            <artifactId>postgresql</artifactId>
-            <version>42.7.3</version>
-        </dependency>
-        <dependency>
-            <groupId>org.testcontainers</groupId>
-            <artifactId>postgresql</artifactId>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>persistence-api</artifactId>
+            <artifactId>datarecords</artifactId>
             <version>${project.version}</version>
         </dependency>
     </dependencies>
-    <build>
+    <properties>
+        <exec.mainClass>io.github.fontysvenlo.ais.persistence.api.PersistenceApi</exec.mainClass>
+    </properties>
+        <build>
         <plugins>
             <plugin>
                 <artifactId>maven-surefire-plugin</artifactId>
