+++ backend/persistence-api/src/main/java/module-info.java
@@ -1,0 +1,9 @@
+/**
+ * Module persistence_api_module.
+ */
+module persistence_api_module { 
+    
+    requires datarecords_module;
+    
+    exports io.github.fontysvenlo.ais.persistence.api;
+}
