+++ backend/persistence/src/main/java/io/github/fontysvenlo/ais/persistence/CustomerRepositoryImpl.java
@@ -1,5 +1,6 @@
 package io.github.fontysvenlo.ais.persistence;
 
+import io.github.fontysvenlo.ais.persistence.api.CustomerRepository;
 import java.time.LocalDate;
 import java.util.List;
 import javax.sql.DataSource;
