+++ backend/persistence/src/main/java/io/github/fontysvenlo/ais/persistence/PersistenceAPIImpl.java
@@ -1,5 +1,8 @@
 package io.github.fontysvenlo.ais.persistence;
 
+import io.github.fontysvenlo.ais.persistence.api.CustomerRepository;
+import io.github.fontysvenlo.ais.persistence.api.PersistenceAPI;
+
 /**
  * Actual creator of storage services.
  */
