+++ backend/persistence/src/main/java/module-info.java
@@ -3,6 +3,7 @@
  */
 module persistence_module { 
     requires datarecords_module;
+    requires persistence_api_module;
     
     requires java.logging;
     requires java.sql;
